requirements.txt
streamlit
google-genai



